+++ moviecatalog/src/main/java/com/drbsimon/moviecatalog/controller/ShowController.java
@@ -1,6 +1,6 @@
 package com.drbsimon.moviecatalog.controller;
 
-import com.drbsimon.moviecatalog.entity.Show;
+import com.drbsimon.moviecatalog.model.Show;
 import com.drbsimon.moviecatalog.model.wrapper.MovieListWrapper;
 import com.drbsimon.moviecatalog.model.wrapper.ShowListWrapper;
 import com.drbsimon.moviecatalog.repository.ShowManager;
